# internal/restic: Don't allocate in Tree.Insert